Move tests for domain specific Repositories
