+++ src/test/java/org/cbioportal/infrastructure/repository/clickhouse/clinical_event/ClickhouseClinicalEventMapperTest.java
@@ -1,7 +1,9 @@
-package org.cbioportal.legacy.persistence.mybatisclickhouse;
+package org.cbioportal.infrastructure.repository.clickhouse.clinical_event;
 
-import org.cbioportal.legacy.persistence.helper.StudyViewFilterHelper;
-import org.cbioportal.legacy.persistence.mybatisclickhouse.config.MyBatisConfig;
+import org.cbioportal.domain.studyview.StudyViewFilterContext;
+import org.cbioportal.domain.studyview.StudyViewFilterFactory;
+import org.cbioportal.infrastructure.repository.clickhouse.AbstractTestcontainers;
+import org.cbioportal.infrastructure.repository.clickhouse.config.MyBatisConfig;
 import org.cbioportal.legacy.web.parameter.DataFilter;
 import org.cbioportal.legacy.web.parameter.DataFilterValue;
 import org.cbioportal.legacy.web.parameter.StudyViewFilter;
@@ -13,39 +15,37 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
 
 import java.util.List;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
-
+import static org.junit.Assert.*;
 @RunWith(SpringRunner.class)
 @Import(MyBatisConfig.class)
 @DataJpaTest
 @DirtiesContext
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 @ContextConfiguration(initializers = AbstractTestcontainers.Initializer.class)
-public class ClinicalEventTypeCountsTest extends AbstractTestcontainers {
-
+public class ClickhouseClinicalEventMapperTest {
     private static final String STUDY_TCGA_PUB = "study_tcga_pub";
 
     @Autowired
-    private StudyViewMapper studyViewMapper;
+    private ClickhouseClinicalEventMapper mapper;
 
     @Test
     public void getClinicalEventTypeCounts() {
         StudyViewFilter studyViewFilter = new StudyViewFilter();
         studyViewFilter.setStudyIds(List.of(STUDY_TCGA_PUB));
 
-        var clinicalEventTypeCounts = studyViewMapper.getClinicalEventTypeCounts(StudyViewFilterHelper.build(studyViewFilter, null, null, studyViewFilter.getStudyIds()));
+        StudyViewFilterContext studyViewFilterContext = StudyViewFilterFactory.make(studyViewFilter, null,
+                studyViewFilter.getStudyIds(), null );
+
+        var clinicalEventTypeCounts = mapper.getClinicalEventTypeCounts(studyViewFilterContext);
 
         assertEquals(4, clinicalEventTypeCounts.size());
 
         var clinicalEventTypeCountOptional = clinicalEventTypeCounts.stream().filter(ce -> ce.getEventType().equals("Treatment"))
-            .findFirst();
+                .findFirst();
 
         assertTrue(clinicalEventTypeCountOptional.isPresent());
         assertEquals(1, clinicalEventTypeCountOptional.get().getCount().intValue());
@@ -56,12 +56,14 @@ public class ClinicalEventTypeCountsTest extends AbstractTestcontainers {
         dataFilter.setValues(List.of(dataFilterValue));
         studyViewFilter.setClinicalEventFilters(List.of(dataFilter));
 
-        clinicalEventTypeCounts = studyViewMapper.getClinicalEventTypeCounts(StudyViewFilterHelper.build(studyViewFilter, null, null, studyViewFilter.getStudyIds()));
+        clinicalEventTypeCounts = mapper.getClinicalEventTypeCounts(
+                StudyViewFilterFactory.make(studyViewFilter, null,
+                        studyViewFilter.getStudyIds(), null ));
 
         assertEquals(3, clinicalEventTypeCounts.size());
 
         clinicalEventTypeCountOptional = clinicalEventTypeCounts.stream().filter(ce -> ce.getEventType().equals("status"))
-            .findFirst();
+                .findFirst();
 
         assertFalse(clinicalEventTypeCountOptional.isPresent());
     }
